requirements.txt
fastapi
transformers
uvicorn
python-multipart
torch
torchvision
torchaudio
ffmpeg
demucs @ git+https://github.com/facebookresearch/demucs#egg=demucs
websockets
google-cloud-storage